backend/requirements.txt
# requirements.txt
fastapi>=0.115.4
uvicorn>=0.32.0
python-dotenv>=1.0.1
web3>=6.11.3
pandas>=2.2.3
scikit-learn>=1.5.2
numpy
pyjwt
aiohttp
python-multipart
pydantic>=2.9.2
